refactor for api
